Add PirTestUtils to _TestUtilities
