+++ Tests/PrivateInformationRetrievalTests/IndexPirTests.swift
@@ -14,7 +14,7 @@
 
 import _TestUtilities
 import HomomorphicEncryption
-@testable import PrivateInformationRetrieval
+import PrivateInformationRetrieval
 import Testing
 
 @Suite
@@ -30,7 +30,7 @@ struct IndexPirTests {
                                             batchSize: 1,
                                             unevenDimensions: false,
                                             keyCompression: .noCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             #expect(parameter.dimensions == [4, 4])
         }
         do {
@@ -40,7 +40,7 @@ struct IndexPirTests {
                                             batchSize: 2,
                                             unevenDimensions: false,
                                             keyCompression: .noCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             #expect(parameter.dimensions == [4, 3])
         }
         // unevenDimensions: true
@@ -51,7 +51,7 @@ struct IndexPirTests {
                                             batchSize: 1,
                                             unevenDimensions: true,
                                             keyCompression: .noCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             #expect(parameter.dimensions == [5, 3])
         }
         do {
@@ -61,7 +61,7 @@ struct IndexPirTests {
                                             batchSize: 2,
                                             unevenDimensions: true,
                                             keyCompression: .noCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             #expect(parameter.dimensions == [5, 3])
         }
         do {
@@ -71,7 +71,7 @@ struct IndexPirTests {
                                             batchSize: 2,
                                             unevenDimensions: true,
                                             keyCompression: .noCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             #expect(parameter.dimensions == [9, 2])
         }
         // no key compression
@@ -82,7 +82,7 @@ struct IndexPirTests {
                                             batchSize: 2,
                                             unevenDimensions: true,
                                             keyCompression: .noCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             let evalKeyConfig = EvaluationKeyConfig(
                 galoisElements: [3, 5, 9, 17],
                 hasRelinearizationKey: true)
@@ -96,7 +96,7 @@ struct IndexPirTests {
                                             batchSize: 2,
                                             unevenDimensions: true,
                                             keyCompression: .hybridCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             let evalKeyConfig = EvaluationKeyConfig(
                 galoisElements: [3, 5, 9, 17],
                 hasRelinearizationKey: true)
@@ -110,7 +110,7 @@ struct IndexPirTests {
                                             batchSize: 2,
                                             unevenDimensions: true,
                                             keyCompression: .maxCompression)
-            let parameter = MulPir.generateParameter(config: config, with: context)
+            let parameter = MulPir<Bfv<UInt64>>.generateParameter(config: config, with: context)
             let evalKeyConfig = EvaluationKeyConfig(
                 galoisElements: [3, 5, 9],
                 hasRelinearizationKey: true)
@@ -118,107 +118,10 @@ struct IndexPirTests {
         }
     }
 
-    private func getDatabaseForTesting(numberOfEntries: Int,
-                                       entrySizeInBytes: Int) -> [[UInt8]]
-    {
-        (0..<numberOfEntries).map { _ in (0..<entrySizeInBytes)
-            .map { _ in UInt8.random(in: UInt8.min...UInt8.max) }
-        }
-    }
-
-    private func indexPirTestForParameter<Server: IndexPirServer, Client: IndexPirClient>(
-        server _: Server.Type,
-        client _: Client.Type,
-        for parameter: IndexPirParameter,
-        with context: Context<Server.Scheme>) throws
-        where Server.IndexPir == Client.IndexPir
-    {
-        let database = getDatabaseForTesting(
-            numberOfEntries: parameter.entryCount,
-            entrySizeInBytes: parameter.entrySizeInBytes)
-        let processedDb = try Server.process(database: database, with: context, using: parameter)
-
-        let server = try Server(parameter: parameter, context: context, database: processedDb)
-        let client = Client(parameter: parameter, context: context)
-
-        let secretKey = try context.generateSecretKey()
-        let evaluationKey = try client.generateEvaluationKey(using: secretKey)
-
-        for _ in 0..<10 {
-            var indices = Array(0..<parameter.batchSize)
-            indices.shuffle()
-            let batchSize = Int.random(in: 1...parameter.batchSize)
-            let queryIndices = Array(indices.prefix(batchSize))
-            let query = try client.generateQuery(at: queryIndices, using: secretKey)
-            let response = try server.computeResponse(to: query, using: evaluationKey)
-            if Server.Scheme.self != NoOpScheme.self {
-                #expect(!response.isTransparent())
-            }
-            let decryptedResponse = try client.decrypt(response: response, at: queryIndices, using: secretKey)
-            for index in queryIndices.indices {
-                #expect(decryptedResponse[index] == database[queryIndices[index]])
-            }
-        }
-    }
-
-    private func indexPirTest<Server: IndexPirServer, Client: IndexPirClient>(server: Server.Type,
-                                                                              client: Client.Type) throws
-        where Server.IndexPir == Client.IndexPir
-    {
-        let configs = try [
-            IndexPirConfig(entryCount: 100,
-                           entrySizeInBytes: 1,
-                           dimensionCount: 2,
-                           batchSize: 2,
-                           unevenDimensions: false,
-                           keyCompression: .noCompression),
-            IndexPirConfig(entryCount: 100,
-                           entrySizeInBytes: 8,
-                           dimensionCount: 2,
-                           batchSize: 2,
-                           unevenDimensions: false,
-                           keyCompression: .noCompression),
-            IndexPirConfig(entryCount: 100,
-                           entrySizeInBytes: 24,
-                           dimensionCount: 2,
-                           batchSize: 2,
-                           unevenDimensions: true,
-                           keyCompression: .noCompression),
-            IndexPirConfig(entryCount: 100,
-                           entrySizeInBytes: 24,
-                           dimensionCount: 1,
-                           batchSize: 2,
-                           unevenDimensions: true,
-                           keyCompression: .noCompression),
-            IndexPirConfig(entryCount: 100,
-                           entrySizeInBytes: 24,
-                           dimensionCount: 1,
-                           batchSize: 2,
-                           unevenDimensions: true,
-                           keyCompression: .hybridCompression),
-            IndexPirConfig(entryCount: 100,
-                           entrySizeInBytes: 24,
-                           dimensionCount: 1,
-                           batchSize: 2,
-                           unevenDimensions: true,
-                           keyCompression: .maxCompression),
-        ]
-
-        let context: Context<Server.Scheme> = try TestUtils.getTestContext()
-        for config in configs {
-            let parameter = Server.generateParameter(config: config, with: context)
-            try indexPirTestForParameter(server: server, client: client, for: parameter, with: context)
-        }
-    }
-
-    private func mulIndexPirTest<Scheme: HeScheme>(scheme _: Scheme.Type) throws {
-        try indexPirTest(server: MulPirServer<Scheme>.self, client: MulPirClient<Scheme>.self)
-    }
-
     @Test
     func indexPir() throws {
-        try mulIndexPirTest(scheme: NoOpScheme.self)
-        try mulIndexPirTest(scheme: Bfv<UInt32>.self)
-        try mulIndexPirTest(scheme: Bfv<UInt64>.self)
+        try PirTestUtils.IndexPirTests.indexPir(scheme: NoOpScheme.self)
+        try PirTestUtils.IndexPirTests.indexPir(scheme: Bfv<UInt32>.self)
+        try PirTestUtils.IndexPirTests.indexPir(scheme: Bfv<UInt64>.self)
     }
 }
